+++ Common/src/main/java/fuzs/puzzleslib/impl/init/RegistryManagerImpl.java
@@ -6,10 +6,14 @@ import fuzs.puzzleslib.api.core.v1.ModLoaderEnvironment;
 import fuzs.puzzleslib.api.core.v1.utility.ResourceLocationHelper;
 import fuzs.puzzleslib.api.init.v3.registry.RegistryHelper;
 import fuzs.puzzleslib.api.init.v3.registry.RegistryManager;
+import fuzs.puzzleslib.impl.item.CreativeModeTabHelper;
 import net.minecraft.core.Holder;
 import net.minecraft.core.Registry;
+import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.item.CreativeModeTab;
+import net.minecraft.world.item.ItemStack;
 import org.apache.commons.lang3.StringUtils;
 
 import java.util.Arrays;
@@ -73,4 +77,18 @@ public abstract class RegistryManagerImpl implements RegistryManager {
     }
 
     protected abstract <T> Holder.Reference<T> getHolderReference(ResourceKey<? extends Registry<? super T>> registryKey, String path, Supplier<T> supplier, boolean skipRegistration);
+
+    @Override
+    public Holder.Reference<CreativeModeTab> registerCreativeModeTab(String path, Supplier<ItemStack> iconSupplier, CreativeModeTab.DisplayItemsGenerator displayItems, boolean withSearchBar) {
+        return this.register(Registries.CREATIVE_MODE_TAB, path, () -> {
+            CreativeModeTab.Builder builder = this.getCreativeModeTabBuilder(withSearchBar);
+            ResourceLocation resourceLocation = this.makeKey(path);
+            builder.title(CreativeModeTabHelper.getTitle(resourceLocation));
+            builder.icon(iconSupplier);
+            builder.displayItems(displayItems);
+            return builder.build();
+        });
+    }
+
+    protected abstract CreativeModeTab.Builder getCreativeModeTabBuilder(boolean withSearchBar);
 }