+++ Fabric/src/main/java/fuzs/puzzleslib/fabric/impl/init/FabricRegistryManager.java
@@ -8,6 +8,7 @@ import fuzs.puzzleslib.api.network.v3.codec.ExtraStreamCodecs;
 import fuzs.puzzleslib.impl.init.DirectReferenceHolder;
 import fuzs.puzzleslib.impl.init.RegistryManagerImpl;
 import net.fabricmc.fabric.api.command.v2.ArgumentTypeRegistry;
+import net.fabricmc.fabric.api.itemgroup.v1.FabricItemGroup;
 import net.fabricmc.fabric.api.object.builder.v1.world.poi.PointOfInterestHelper;
 import net.fabricmc.fabric.api.screenhandler.v1.ExtendedScreenHandlerType;
 import net.minecraft.commands.synchronization.ArgumentTypeInfo;
@@ -22,6 +23,7 @@ import net.minecraft.world.entity.Mob;
 import net.minecraft.world.entity.ai.village.poi.PoiType;
 import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.inventory.MenuType;
+import net.minecraft.world.item.CreativeModeTab;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.SpawnEggItem;
 import net.minecraft.world.level.block.state.BlockState;
@@ -54,21 +56,31 @@ public final class FabricRegistryManager extends RegistryManagerImpl {
     }
 
     @Override
+    protected CreativeModeTab.Builder getCreativeModeTabBuilder(boolean withSearchBar) {
+        return FabricItemGroup.builder();
+    }
+
+    @Override
     public Holder.Reference<Item> registerSpawnEggItem(Holder<? extends EntityType<? extends Mob>> entityTypeReference, int backgroundColor, int highlightColor, Item.Properties itemProperties) {
-        return this.registerItem(entityTypeReference.unwrapKey().orElseThrow().location().getPath() + "_spawn_egg", () -> new SpawnEggItem(entityTypeReference.value(), backgroundColor, highlightColor, itemProperties));
+        return this.registerItem(entityTypeReference.unwrapKey().orElseThrow().location().getPath() + "_spawn_egg",
+                () -> new SpawnEggItem(entityTypeReference.value(), backgroundColor, highlightColor, itemProperties));
     }
 
     @SuppressWarnings("unchecked")
     @Override
     public <T extends AbstractContainerMenu> Holder.Reference<MenuType<T>> registerExtendedMenuType(String path, Supplier<ExtendedMenuSupplier<T>> entry) {
-        return this.register((ResourceKey<Registry<MenuType<T>>>) (ResourceKey<?>) Registries.MENU, path, () -> new ExtendedScreenHandlerType<>(entry.get()::create,
-                ExtraStreamCodecs.REGISTRY_FRIENDLY_BYTE_BUF
-        ));
+        return this.register((ResourceKey<Registry<MenuType<T>>>) (ResourceKey<?>) Registries.MENU,
+                path,
+                () -> new ExtendedScreenHandlerType<>(entry.get()::create,
+                        ExtraStreamCodecs.REGISTRY_FRIENDLY_BYTE_BUF));
     }
 
     @Override
     public Holder.Reference<PoiType> registerPoiType(String path, Supplier<Set<BlockState>> matchingStates, int maxTickets, int validRange) {
-        return this.register(Registries.POINT_OF_INTEREST_TYPE, path, () -> PointOfInterestHelper.register(this.makeKey(path), maxTickets, validRange, matchingStates.get()), true);
+        return this.register(Registries.POINT_OF_INTEREST_TYPE,
+                path,
+                () -> PointOfInterestHelper.register(this.makeKey(path), maxTickets, validRange, matchingStates.get()),
+                true);
     }
 
     @Override
@@ -82,8 +94,7 @@ public final class FabricRegistryManager extends RegistryManagerImpl {
     @Override
     public <T> Holder.Reference<EntityDataSerializer<T>> registerEntityDataSerializer(String path, Supplier<EntityDataSerializer<T>> entry) {
         ResourceKey<Registry<EntityDataSerializer<?>>> registryKey = ResourceKey.createRegistryKey(
-                ResourceLocationHelper.withDefaultNamespace(
-                "entity_data_serializers"));
+                ResourceLocationHelper.withDefaultNamespace("entity_data_serializers"));
         EntityDataSerializer<T> serializer = entry.get();
         EntityDataSerializers.registerSerializer(serializer);
         return new DirectReferenceHolder<>(this.makeResourceKey(registryKey, path), serializer);