+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.1.28-1.21.1] - 2025-02-25
+### Added
+- Backport new block, item and creative mode tab registration methods to `RegistryManager`
+- Backport `CompostableBlocksContext`
+### Changed
+- Update `FabricEventFactory` to handle faulty return values more leniently
+
 ## [v21.1.27-1.21.1] - 2025-01-17
 ### Fixed
 - Fix network packets executing with one tick delay on Fabric